+++ compiler/repl-service/server/src/main/scala/com/digitalasset/daml/lf/speedy/ReplServiceMain.scala
@@ -26,7 +26,8 @@ import java.net.{InetAddress, InetSocketAddress}
 import java.nio.file.{Files, Path, Paths}
 import java.util.logging.{Level, Logger}
 
-import com.daml.lf.engine.script.ledgerinteraction.{ScriptLedgerClient, ScriptTimeMode}
+import com.daml.lf.engine.script.ScriptTimeMode
+import com.daml.lf.engine.script.ledgerinteraction.ScriptLedgerClient
 import com.daml.lf.speedy.iterable.SExprIterable
 
 import scala.jdk.CollectionConverters._
@@ -299,7 +300,7 @@ class ReplService(
           val e = originalE match {
             // For now, don’t show stack traces in Daml Repl. They look fairly confusing
             // since they refer to the internal names we use.
-            case e: ScriptF.FailedCmd => e.cause
+            case e: Script.FailedCmd => e.cause
             case _ => originalE
           }
           respObs.onNext(
