+++ daml-script/runner/src/main/scala/com/digitalasset/daml/lf/engine/script/v1/ledgerinteraction/IdeLedgerClient.scala
@@ -4,6 +4,7 @@
 package com.daml.lf
 package engine
 package script
+package v1
 package ledgerinteraction
 
 import akka.stream.Materializer
